third_party/blink/web_tests/platform/generic/fast/css/counters/counter-cssText-expected.txt

Test the cssText output of counter-valued CSSPrimitiveValues

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS rules[2].style.cssText is "content: counter(section);"
PASS rules[3].style.cssText is "content: counters(section, \":\");"
PASS rules[4].style.cssText is "content: counter(section, lower-roman);"
PASS rules[5].style.cssText is "content: counters(section, \",\", upper-roman);"
PASS rules[6].style.cssText is "content: counter(section, none);"
PASS successfullyParsed is true

TEST COMPLETE
